Working so far.

+++ Helpers.cuh
@@ -9,13 +9,13 @@
 
 #include "stdio.h"
 
-__device__ inline int32_t GetZone(const int32_t y);
+__device__ inline uint8_t GetZone(const uint8_t y);
 __device__ inline void CreateDoor(bbRandom* bb, rnd_state* rnd_state, bool open, int32_t big);
 __device__ inline void CreateItem(bbRandom* bb, rnd_state* rnd_state);
 __host__ __device__ inline char* RoomIDToName(RoomID r);
 
-__device__ inline int32_t GetZone(const int32_t y) {
-	int32_t zone = int32_t(min(floorf((float(MapWidth - y) / MapWidth * ZONEAMOUNT)), float(ZONEAMOUNT - 1)));
+__device__ inline uint8_t GetZone(const uint8_t y) {
+	uint8_t zone = uint8_t(min(floorf((float(MapWidth - y) / MapWidth * ZONEAMOUNT)), float(ZONEAMOUNT - 1)));
 	return zone;
 }
